Fixed problem where image-path was absolute, not relative.
Fixed position of click-able area for damage button. The location of the button is relative to `scrollable_rect`/`scrollable_content`, but the mouse-position is absolute, so the rect of the buttons has to be moved by the x and y postion of `scrollable_rect`. In addition, the y position of `scrollable_rect` has to be subtracted by `target_scroll_position` to account for the scroll.